It's time to ask for donations.
For the next year I'm going to be in Tanzania w/ my girlfriend who is joining Peace Corps Response and finances are going to be fairly tight. I don't like asking for money, because I earn a really good salary when I'm in the US, but frankly, Tanzania is more complicated financially and living in Dar is not exactly cheap. And if accepting tips means the difference between keeping the Internet on...

I'm not expecting much, but every little bit will help.